Changed some stuff for 64 bit support (floats to CGFloat, etc.)

+++ Framework/Views/OpenGL/KTOpenGLControl.m
@@ -75,7 +75,7 @@
 //=========================================================== 
 // - setFloatValue
 //===========================================================
-- (void)setFloatValue:(float)theFloatValue
+- (void)setFloatValue:(CGFloat)theFloatValue
 {
 	mFloatValue = theFloatValue;
 	[self setNeedsDisplay:YES];
